refactor : user 서비스 클래스, 인터페이스를 user 패키지 내부로 이동함
