+++ src/main/java/kr/ac/cnuswacademy/springcoffeeshop/controller/UserController.java
@@ -3,13 +3,12 @@ package kr.ac.cnuswacademy.springcoffeeshop.controller;
 import kr.ac.cnuswacademy.springcoffeeshop.dto.user.UserListResponseDto;
 import kr.ac.cnuswacademy.springcoffeeshop.dto.user.UserResponseDto;
 import kr.ac.cnuswacademy.springcoffeeshop.dto.user.UserSaveRequestDto;
-import kr.ac.cnuswacademy.springcoffeeshop.service.UserService;
+import kr.ac.cnuswacademy.springcoffeeshop.service.user.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.List;
